src/aiservices/requirements.txt
flask
requests
gunicorn
google-cloud-aiplatform
mcp
